Change uniform_int_distribution type
Some implementations don't allow single-byte types
https://stackoverflow.com/questions/31460733/why-arent-stduniform-int-distributionuint8-t-and-stduniform-int-distri

+++ src/tests/test_runtime.cpp
@@ -679,7 +679,7 @@ TEST_CASE("CBVar-comparison", "[ops]") {
 
   SECTION("Bytes") {
     std::random_device rd;
-    std::uniform_int_distribution<uint8_t> dist(0, 0x07);
+    std::uniform_int_distribution<uint32_t> dist(0, 0x07);
 
     std::vector<uint8_t> data1(1024);
     for (auto &b : data1) {
